+++ PSServicePrincipal/functions/Get-AppAndSPNPair.ps1
@@ -47,6 +47,6 @@
     }
     catch
     {
-        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_
+        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_ -EnableException $true
     }
 }