+++ PSServicePrincipal/functions/Get-EnterpriseApp.ps1
@@ -31,6 +31,6 @@
     }
     catch
     {
-        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_
+        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_ -EnableException $true
     }
 }