+++ PSServicePrincipal/functions/Get-SpnByAppID.ps1
@@ -49,6 +49,6 @@
     }
     catch
     {
-        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_
+        Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_ -EnableException
     }
 }